Create components for forms
Move logic and rendering for word forms out of page functions and into individual components.
